Moved Token class to own file
